+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Get-ClassResourceProperty`
   - Now does a more limited wildcard search for the class script file ([issue #122](https://github.com/dsccommunity/DscResource.DocGenerator/issues/122)).
+  - Regression tests.
 
 ## [0.11.0] - 2022-05-10